requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
black==24.10.0
bleach==6.2.0
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.3
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.7.0
docutils==0.21.2
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
gotrue==2.9.3
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
isort==5.13.2
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.0
jupyterlab_code_formatter==3.0.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
lxml==5.3.0
MAPIE==0.9.1
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nicegui==2.5.0
notebook_shim==0.2.4
numpy==1.26.4
orjson==3.10.11
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
playwright==1.48.0
plotly==5.24.1
postgrest==0.17.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
Protego==0.3.1
pscript==0.7.7
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
pyee==12.0.0
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==2.0.7
python-multipart==0.0.17
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
realtime==2.0.6
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.11.2
scrapy-playwright==0.0.41
scrapy-user-agents==0.1.1
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==75.3.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.2
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
Twisted==24.10.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
uri-template==1.3.0
urllib3==2.2.3
user-agents==2.2.0
uvicorn==0.32.0
uvloop==0.21.0
vbuild==0.8.2
w3lib==2.2.1
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
yarl==1.17.0
zope.interface==7.1.1
